requirements.txt
streamlit==1.38.0
ultralytics==8.2.97
pillow==10.4.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
firebase-admin==6.5.0
bcrypt==4.1.3
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.140.0

